Add arguments from RUSTFLAGS to the rustc command.
This fixes feature probing in cross-compilation environments, for example using Yocto and its meta-rust layer.
RUSTFLAGS is supported by cargo to pass flags to all rustc invocations. Yocto uses it to pass the library search path of the target sysroot, which is required to compile anything for the given target.

I'm using the shlex crate to properly parse arguments in RUSTFLAGS, also those containing whitespaces.