Support `RUSTC_WRAPPER` and `RUSTC_WORKSPACE_WRAPPER`
We need Cargo 1.55 to know whether the workspace wrapper is applicable,
and even `RUSTC_WRAPPER` is incomplete before then because it could also
be set in config files. The environment does take precedence when set
though, so it should hopefully be fine to use it on earlier versions.
The status quo remains to miss config wrappers before Rust 1.55.